added calorie overrides
