+++ embed.go
@@ -2,5 +2,5 @@ package fitness
 
 import "embed"
 
-//go:embed templates/index.tmpl
+//go:embed templates/index.html
 var Content embed.FS
