+++ templates/index.html
@@ -1,4 +1,4 @@
-{{ define "index.tmpl" }}
+{{ define "index.html" }}
 <html lang="en">
 
 <head>
